+++ NEWS.md
@@ -1,3 +1,16 @@
+# patternogram 0.1.1
+
+* Adds parallalization possibility when calculating uncertainty intervals
+
+# patternogram 0.1.0
+
+* Allows to calculate two types of confidence intervals for patternograms
+* Renames the "target" argument to "group" for clarity
+
+# patternogram 0.0.1
+
+* Initial working version of the package as presented at ECEM 2023
+
 # patternogram 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.